+++ UD4SpringGrupo1/src/main/java/com/example/demo/model/ServicioModel.java
@@ -30,7 +30,7 @@ public class ServicioModel {
 
 	 private ProFamily profesionalFamilyId;
 	
-	private int valoration;
+	private float valoration;
 	
 	private int finished;
 	
@@ -86,10 +86,10 @@ public class ServicioModel {
 	public void setProfesionalFamilyId(ProFamily profesionalFamilyId) {
 		this.profesionalFamilyId = profesionalFamilyId;
 	}
-	public int getValoration() {
+	public float getValoration() {
 		return valoration;
 	}
-	public void setValoration(int valoration) {
+	public void setValoration(float valoration) {
 		this.valoration = valoration;
 	}
 	public int isFinished() {
@@ -111,7 +111,7 @@ public class ServicioModel {
 		this.deleted = deleted;
 	}
 	public ServicioModel(int id, String title, String description, Date registerDate, Date happeningDate, Student studentId,
-			Business businessId, ProFamily profesionalFamilyId, int valoration, int finished, String comment, int deleted) {
+			Business businessId, ProFamily profesionalFamilyId, float valoration, int finished, String comment, int deleted) {
 		super();
 		this.id = id;
 		this.title = title;
